+++ packages/core/src/editing/rest-component-layout-service.test.ts
@@ -1,24 +1,25 @@
 /* eslint-disable no-unused-expressions */
-import { expect, spy, use } from 'chai';
+import { expect, use } from 'chai';
 import spies from 'chai-spies';
-import { IncomingMessage, ServerResponse } from 'http';
+import nock from 'nock';
+import { NativeDataFetcherConfig } from '../native-fetcher';
 import {
   ComponentLayoutRequestParams,
   RestComponentLayoutService,
 } from './rest-component-layout-service';
-import { EditMode, LayoutServiceData } from '../layout/models';
-import nock from 'nock';
-import { NativeDataFetcher } from '../native-fetcher';
+import { LayoutServiceData } from '../layout/models';
+import { SITECORE_EDGE_URL_DEFAULT } from '../constants';
 
 use(spies);
 
 describe('RestComponentLayoutService', () => {
-  type SetHeader = (name: string, value: unknown) => void;
-
   const defaultTestInput: ComponentLayoutRequestParams = {
     itemId: '123',
     componentUid: '456',
+    siteName: 'supersite',
   };
+
+  const contextId = 'test-context-id';
 
   const defaultTestData = {
     sitecore: {
@@ -36,38 +37,31 @@ describe('RestComponentLayoutService', () => {
     nock.cleanAll();
   });
 
-  it('should fetch component data', async () => {
-    nock('http://sctest')
+  it('should fetch component data', () => {
+    nock(SITECORE_EDGE_URL_DEFAULT)
       .get(
-        '/sitecore/api/layout/component/jss?sc_apikey=0FBFF61E-267A-43E3-9252-B77E71CEE4BA&item=123&uid=456&sc_site=supersite&sc_lang=en'
+        '/layout/component?sitecoreContextId=test-context-id&item=123&uid=456&sc_site=supersite&sc_lang=en'
       )
-      .reply(200, (_) => defaultTestData);
+      .reply(200, () => defaultTestData);
 
     const service = new RestComponentLayoutService({
-      apiHost: 'http://sctest',
-      apiKey: '0FBFF61E-267A-43E3-9252-B77E71CEE4BA',
-      siteName: 'supersite',
+      contextId,
     });
-    const result = await service.fetchComponentData(defaultTestInput);
-    expect(result).to.deep.equal(defaultTestData);
+
+    return service
+      .fetchComponentData(defaultTestInput)
+      .then((layoutServiceData: LayoutServiceData & NativeDataFetcherConfig) => {
+        expect(layoutServiceData).to.deep.equal(defaultTestData);
+      });
   });
 
   it('should fetch component data when optional params provided', () => {
     const testInput: ComponentLayoutRequestParams = {
       ...defaultTestInput,
       dataSourceId: '789',
-    };
-
-    const testUnexpectedData = {
-      sitecore: {
-        context: {},
-        route: {
-          name: 'xxx',
-          placeholders: {
-            'editing-componentmode-placeholder': [],
-          },
-        },
-      },
+      renderingId: '000',
+      version: '1',
+      language: 'en',
     };
 
     const testExpectedData = {
@@ -95,182 +89,48 @@ describe('RestComponentLayoutService', () => {
       },
     };
 
-    nock('http://sctest')
+    nock(SITECORE_EDGE_URL_DEFAULT)
       .get(
-        '/sitecore/api/layout/component/jss?sc_apikey=0FBFF61E-267A-43E3-9252-B77E71CEE4BA&item=123&uid=456&dataSourceId=789&sc_site=supersite&sc_lang=en'
+        '/layout/component?sitecoreContextId=test-context-id&item=123&uid=456&dataSourceId=789&renderingItemId=000&version=1&sc_site=supersite&sc_lang=en'
       )
-      .reply(200, (_) => testExpectedData)
-      .get('/sitecore/api/layout/component/jss')
-      .query(true)
-      .reply(200, (_) => testUnexpectedData);
-
-    const req = {
-      connection: {
-        remoteAddress: '192.168.1.10',
-      },
-      headers: {
-        cookie: 'test-cookie-value',
-        referer: 'http://sctest',
-        'user-agent': 'test-user-agent-value',
-      },
-    } as IncomingMessage;
-
-    const setHeaderSpy: SetHeader = spy();
-
-    const res = {
-      setHeader: setHeaderSpy,
-    } as ServerResponse;
+      .reply(200, () => testExpectedData);
 
     const service = new RestComponentLayoutService({
-      apiHost: 'http://sctest',
-      apiKey: '0FBFF61E-267A-43E3-9252-B77E71CEE4BA',
-      siteName: 'supersite',
+      contextId,
     });
 
     return service
-      .fetchComponentData(testInput, req, res)
-      .then((layoutServiceData: LayoutServiceData) => {
+      .fetchComponentData(testInput)
+      .then((layoutServiceData: LayoutServiceData & NativeDataFetcherConfig) => {
         expect(layoutServiceData).to.deep.equal(testExpectedData);
       });
   });
 
-  it('should fetch component data with custom site name', () => {
-    const testInput: ComponentLayoutRequestParams = {
-      ...defaultTestInput,
-      siteName: 'mysite',
-    };
-
-    const testUnexpectedData = {
-      sitecore: {
-        context: {},
-        route: {
-          name: 'xxx',
-          placeholders: {
-            'editing-componentmode-placeholder': [],
-          },
-        },
-      },
-    };
-
-    const testExpectedData = {
-      sitecore: {
-        context: {},
-        route: {
-          name: 'xxx',
-          placeholders: {
-            'editing-componentmode-placeholder': [
-              {
-                uid: '456',
-                componentName: 'RichText',
-                dataSource: '789',
-                params: {
-                  GridParameters: 'col-12',
-                  FieldNames: 'Default',
-                  Styles: '',
-                  RenderingIdentifier: '',
-                  DynamicPlaceholderId: '3',
-                },
-              },
-            ],
-          },
-        },
-      },
-    };
-
-    nock('http://sctest')
-      .get(
-        '/sitecore/api/layout/component/jss?sc_apikey=0FBFF61E-267A-43E3-9252-B77E71CEE4BA&item=123&uid=456&sc_site=mysite&sc_lang=en'
-      )
-      .reply(200, (_) => testExpectedData)
-      .get('/sitecore/api/layout/component/jss')
-      .query(true)
-      .reply(200, (_) => testUnexpectedData);
-
-    const req = {
-      connection: {
-        remoteAddress: '192.168.1.10',
-      },
-      headers: {
-        cookie: 'test-cookie-value',
-        referer: 'http://sctest',
-        'user-agent': 'test-user-agent-value',
-      },
-    } as IncomingMessage;
-
-    const setHeaderSpy: SetHeader = spy();
-
-    const res = {
-      setHeader: setHeaderSpy,
-    } as ServerResponse;
+  it('should fetch component data from a custom edge endpoint', () => {
+    const customEdgeUrl = 'https://custom-edge-url.com';
 
     const service = new RestComponentLayoutService({
-      apiHost: 'http://sctest',
-      apiKey: '0FBFF61E-267A-43E3-9252-B77E71CEE4BA',
-      siteName: 'supersite',
+      contextId,
+      edgeUrl: customEdgeUrl,
     });
 
-    return service
-      .fetchComponentData(testInput, req, res)
-      .then((layoutServiceData: LayoutServiceData) => {
-        expect(layoutServiceData).to.deep.equal(testExpectedData);
-      });
-  });
-
-  it('should fetch layout data using custom configuration name', () => {
-    nock('http://sctest')
+    nock(customEdgeUrl)
       .get(
-        '/sitecore/api/layout/component/listen?sc_apikey=0FBFF61E-267A-43E3-9252-B77E71CEE4BA&item=123&uid=456&sc_site=supersite&sc_lang=en'
-      )
-      .reply(200, (_) => defaultTestData);
-
-    const service = new RestComponentLayoutService({
-      apiHost: 'http://sctest',
-      apiKey: '0FBFF61E-267A-43E3-9252-B77E71CEE4BA',
-      siteName: 'supersite',
-      configurationName: 'listen',
-    });
-
-    return service
-      .fetchComponentData(defaultTestInput)
-      .then((layoutServiceData: LayoutServiceData) => {
-        expect(layoutServiceData).to.deep.equal(defaultTestData);
-      });
-  });
-
-  it('should fetch layout data using custom fetcher resolver', () => {
-    const fetcherSpy = spy((url: string) => {
-      return new NativeDataFetcher().fetch<never>(url);
-    });
-
-    nock('http://sctest')
-      .get(
-        '/sitecore/api/layout/component/jss?sc_apikey=0FBFF61E-267A-43E3-9252-B77E71CEE4BA&item=123&uid=456&sc_site=supersite&sc_lang=en'
+        '/layout/component?sitecoreContextId=test-context-id&item=123&uid=456&sc_site=supersite&sc_lang=en'
       )
       .reply(200, () => defaultTestData);
 
-    const service = new RestComponentLayoutService({
-      apiHost: 'http://sctest',
-      apiKey: '0FBFF61E-267A-43E3-9252-B77E71CEE4BA',
-      siteName: 'supersite',
-      dataFetcherResolver: () => fetcherSpy,
-    });
-
     return service
       .fetchComponentData(defaultTestInput)
-      .then((layoutServiceData: LayoutServiceData) => {
+      .then((layoutServiceData: LayoutServiceData & NativeDataFetcherConfig) => {
         expect(layoutServiceData).to.deep.equal(defaultTestData);
-
-        expect(fetcherSpy).to.be.called.once;
-        expect(fetcherSpy).to.be.called.with(
-          'http://sctest/sitecore/api/layout/component/jss?sc_apikey=0FBFF61E-267A-43E3-9252-B77E71CEE4BA&item=123&uid=456&sc_site=supersite&sc_lang=en'
-        );
       });
   });
 
   it('should catch 404 when request layout data', () => {
-    nock('http://sctest')
+    nock(SITECORE_EDGE_URL_DEFAULT)
       .get(
-        '/sitecore/api/layout/component/jss?sc_apikey=0FBFF61E-267A-43E3-9252-B77E71CEE4BA&item=123&uid=456&sc_site=supersite&sc_lang=en'
+        '/layout/component?sitecoreContextId=test-context-id&item=123&uid=456&sc_site=supersite&sc_lang=en'
       )
       .reply(404, () => ({
         data: {
@@ -279,9 +139,7 @@ describe('RestComponentLayoutService', () => {
       }));
 
     const service = new RestComponentLayoutService({
-      apiHost: 'http://sctest',
-      apiKey: '0FBFF61E-267A-43E3-9252-B77E71CEE4BA',
-      siteName: 'supersite',
+      contextId,
     });
 
     return service
@@ -302,16 +160,14 @@ describe('RestComponentLayoutService', () => {
   });
 
   it('should allow non 404 errors through', () => {
-    nock('http://sctest')
+    nock(SITECORE_EDGE_URL_DEFAULT)
       .get(
-        '/sitecore/api/layout/component/jss?sc_apikey=0FBFF61E-267A-43E3-9252-B77E71CEE4BA&item=123&uid=456&sc_site=supersite&sc_lang=en'
+        '/layout/component?sitecoreContextId=test-context-id&item=123&uid=456&sc_site=supersite&sc_lang=en'
       )
       .reply(401, { message: 'whoops' });
 
     const service = new RestComponentLayoutService({
-      apiHost: 'http://sctest',
-      apiKey: '0FBFF61E-267A-43E3-9252-B77E71CEE4BA',
-      siteName: 'supersite',
+      contextId,
     });
 
     return service.fetchComponentData(defaultTestInput).catch((error) => {
@@ -323,9 +179,7 @@ describe('RestComponentLayoutService', () => {
   describe('getComponentFetchParams', () => {
     it('should return params', () => {
       const service = new RestComponentLayoutService({
-        apiHost: 'http://sctest',
-        apiKey: '0FBFF61E-267A-43E3-9252-B77E71CEE4BA',
-        siteName: 'supersite',
+        contextId,
       });
       const testParams = {
         itemId: '123',
@@ -335,12 +189,10 @@ describe('RestComponentLayoutService', () => {
         version: '1',
         siteName: 'notsupersite',
         language: 'en',
-        editMode: EditMode.Metadata,
-        variant: 'default',
       };
 
       const expectedResult = {
-        sc_apikey: '0FBFF61E-267A-43E3-9252-B77E71CEE4BA',
+        sitecoreContextId: contextId,
         item: testParams.itemId,
         uid: testParams.componentUid,
         dataSourceId: testParams.dataSourceId,
@@ -348,7 +200,6 @@ describe('RestComponentLayoutService', () => {
         version: testParams.version,
         sc_site: testParams.siteName,
         sc_lang: testParams.language,
-        sc_mode: testParams.editMode,
       };
 
       // eslint-disable-next-line dot-notation
@@ -357,29 +208,26 @@ describe('RestComponentLayoutService', () => {
 
     it('should return params with no undefined params', () => {
       const service = new RestComponentLayoutService({
-        apiHost: 'http://sctest',
-        apiKey: '0FBFF61E-267A-43E3-9252-B77E71CEE4BA',
-        siteName: 'supersite',
+        contextId,
       });
+
       const testParams = {
         itemId: '123',
         componentUid: '456',
         dataSourceId: undefined,
         renderingId: '000',
         version: undefined,
-        siteName: undefined,
+        siteName: 'supersite',
         language: 'en',
-        editMode: EditMode.Metadata,
-        variant: 'default',
       };
 
       const expectedResult = {
-        sc_apikey: '0FBFF61E-267A-43E3-9252-B77E71CEE4BA',
+        sitecoreContextId: contextId,
         item: testParams.itemId,
         uid: testParams.componentUid,
         renderingItemId: testParams.renderingId,
         sc_lang: testParams.language,
-        sc_mode: testParams.editMode,
+        sc_site: testParams.siteName,
       };
 
       // eslint-disable-next-line dot-notation